fix: that reschedulebooking screen show confirmdialog instead of bookingform
